+++ definition/teams/openapi/components/parameters/teamId.@.js
@@ -1,0 +1,9 @@
+export default {
+	description: 'The identifier of the specific team.',
+	name: 'teamId',
+	in: 'path',
+	required: true,
+	schema: {
+		type: 'string',
+	},
+}