+++ definition/teams/routes/paths/teams/get.@.js
@@ -1,0 +1,49 @@
+export const summary = 'Get Team List'
+
+export const description = `
+Fetch the list of all teams available to the requesting user.
+`
+
+export const tags = [ 'teams' ]
+
+export const security = [
+	{ cookie: [] },
+	{ api: [] },
+]
+
+export const parameters = [
+	{ $ref: '#/components/parameters/filter' },
+	{ $ref: '#/components/parameters/sort' },
+]
+
+export const responses = {
+	200: {
+		description: 'The filtered list of teams.',
+		content: {
+			'application/json': {
+				schema: {
+					type: 'object',
+					properties: {
+						data: {
+							type: 'array',
+							items: {
+								$ref: '#/components/schemas/team',
+							},
+						},
+					},
+				},
+			},
+		},
+	},
+	default: {
+		$ref: '#/components/responses/error',
+	},
+}
+
+export default async request => {
+	const { teams } = await request.controller.team.list(request)
+	return {
+		status: 200,
+		body: { data: teams },
+	}
+}